+++ src/server/Balance/BalanceServer.ts
@@ -6,7 +6,7 @@ export class BalanceServer {
     list = (node: iNode) => {
         let api: ApiModule = new ApiModule(new Balance());
 
-        api.list(node).then((response) => {
+        api.list({ q: node.request.query['q'] }).then((response) => {
             node.response.end(JSON.stringify(response));
         });
     };